back/requirements.txt
Django==5.2
django-cors-headers
django-allauth
cryptography
django-dotnetid
pyjwt
requests
python-dotenv
